requirement.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.12
django-allauth==0.35.0
django-extensions==2.0.6
django-filter==1.1.0
django-rest==0.0.1
django-rest-auth==0.9.3
djangorestframework==3.8.1
djangorestframework-jwt==1.11.0
drf-pyotp==0.5.6
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
MySQL-python==1.2.5
oauthlib==2.0.7
Pillow==5.1.0
Pygments==2.2.0
PyJWT==1.6.1
pyotp==2.2.6
python-openid==2.2.5
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22